requirements.txt
Django>3.0,<4.0
gunicorn
djangorestframework
psycopg2-binary>=2.8
requests
django-heroku
whitenoise
django-extensions
